+++ apps/ioc-maker-cli/examples/abstract-classes-example/container.gen.ts
@@ -1,0 +1,27 @@
+import { UserRepository } from './implementations/UserRepository';
+
+function createCoreModuleContainer() {
+
+  let userRepository: UserRepository | undefined;
+
+  const getUserRepository = (): UserRepository => {
+    if (!userRepository) {
+      userRepository = new UserRepository();
+    }
+    return userRepository;
+  };
+
+  return {
+        get UserRepository(): UserRepository {
+          return getUserRepository();
+        }
+  };
+}
+
+const coreModuleContainer = createCoreModuleContainer();
+
+export const container = {
+  coreModule: coreModuleContainer
+};
+
+export type Container = typeof container;