+++ apps/ioc-maker-cli/examples/clean-architecture/UserModule.gen.ts
@@ -1,4 +1,3 @@
-import { UserRepository } from './repositories/UserRepository';
 import { GetUserUseCase } from './use-cases/GetUserUseCase';
 import { GetTodosByUserUseCase } from './use-cases/GetTodosByUserUseCase';
 import { DeleteUserUseCase } from './use-cases/DeleteUserUseCase';
@@ -7,6 +6,7 @@ import { GetUserPresenter } from './presenters/GetUserPresenter';
 import { GetTodosByUserPresenter } from './presenters/GetTodosByUserPresenter';
 import { DeleteUserPresenter } from './presenters/DeleteUserPresenter';
 import { CreateUserPresenter } from './presenters/CreateUserPresenter';
+import { UserRepository } from './repositories/UserRepository';
 function createUserModuleContainer() {
   const getUserPresenterFactory = (): GetUserPresenter => new GetUserPresenter();
   const getTodosByUserPresenterFactory = (): GetTodosByUserPresenter => new GetTodosByUserPresenter();
@@ -51,12 +51,6 @@ function createUserModuleContainer() {
   };
 
   return {
-        get UserRepository(): UserRepository {
-          return getUserRepository();
-        },
-        get IUserRepository(): UserRepository {
-          return getUserRepository();
-        },
         get GetUserUseCase(): GetUserUseCase {
           return getGetUserUseCase();
         },
@@ -104,6 +98,12 @@ function createUserModuleContainer() {
         },
         get ICreateUserOutputPort(): CreateUserPresenter {
           return createUserPresenterFactory();
+        },
+        get UserRepository(): UserRepository {
+          return getUserRepository();
+        },
+        get IUserRepository(): UserRepository {
+          return getUserRepository();
         }
   };
 }