+++ apps/docs/CHANGELOG.md
@@ -1,5 +1,11 @@
 # docs
 
+## 0.0.3
+
+### Patch Changes
+
+- feat(abstract-classes): add support for abstract classes and inheritance
+
 ## 0.0.2
 
 ### Patch Changes